Checking

+++ es3/connection.cpp
@@ -384,7 +384,7 @@ public:
 		if (res<0)
 			res | libc_die;
 
-		pre_read_.resize(std::min(25*1024*1024, size-offset));
+		pre_read_.resize(std::min(size_t(25*1024*1024), size-offset));
 		prime_offset_=cur_primed_=0;
 		prime();
 	}
